incremental changes
Changes regarding the error handling; when the server is not initialized with admin privileges, the client will be welcomed with a 500 http internal server error page and the log file will contain further instructions. This was made for security reasons due to the fact that the ARP request(in order to determine if the IP does exist on the network) can not be processed unless the server is initialized with admin privileges.