incremental changes
Server Side improvements; The ARP request is now working ONLY on Linux based servers(Windows serverside may be implemented soon). The script is able to determine wether the requested IP to send the files to is active on the network or not. NOTE: High sensitivity between 2.4 GHz and 5 GHz networks.